+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as Button } from './Button';
 export { default as Tooltip } from './Tooltip';
+export { default as SpinButton } from './SpinButton';
